Extract shared interop request logic to separate class
