+++ osu.Server.Queues.ScoreStatisticsProcessor/Processors/MedalProcessor.cs
@@ -102,7 +102,7 @@ namespace osu.Server.Queues.ScoreStatisticsProcessor.Processors
         private void awardMedal(SoloScore score, Medal medal)
         {
             Console.WriteLine($"Awarding medal {medal.name} to user {score.user_id} (score {score.id})");
-            LegacyDatabaseHelper.RunSharedInteropCommand($"user-achievement/{score.user_id}/{medal.achievement_id}/{score.beatmap_id}", "POST");
+            WebRequestHelper.RunSharedInteropCommand($"user-achievement/{score.user_id}/{medal.achievement_id}/{score.beatmap_id}", "POST");
             MedalAwarded?.Invoke(new AwardedMedal(medal, score));
         }
